+++ inst/@mp/mrdivide.m
@@ -1,0 +1,6 @@
+## SPDX-License-Identifier: BSD-2-Clause
+
+function result = mrdivide (varargin)
+  error ("mplapack:NotImplemented", ...
+         "mp right division is not implemented before M09");
+endfunction