+++ src/mp_scalar_storage.cc
@@ -72,6 +72,11 @@ MpfrScalarStorage::MpfrScalarStorage (double value,
   mpfr_set_d (m_value.mpfr_data (), value, MPFR_RNDN);
 }
 
+MpfrScalarStorage::MpfrScalarStorage (NativeScalar value)
+  : m_value (std::move (value))
+{
+}
+
 MpfrScalarStorage::MpfrScalarStorage (mpfr_prec_t precision_bits,
                                       UninitializedTag)
   : m_value (NativeScalar::with_precision (precision_bits))